requirements.txt
langchain
langchain_community
openai
psycopg2-binary
streamlit
pdfplumber